dbw_node.py: Some formatting changes to avoid lines that are too long.
twist_controller.py:
Code clean up: use named args instead of indexing into args[].
Improve PID controller by setting output range to [-1.0 1.0] so that we
can directly use the non-negative values (in range [0 1.0]) directly as
the throttle control.
For the brake, calculate the max_brake value in torque (Nm) based on the
vehicle_mass, decel_limit and wheel_radius. Then we can scale negative
values of the PID control to generate the desired brake value and can
be assured not to exceed decel_limit.

This results in a smoother ride with throttle and brake values
that are not changing drastically.