Add code to conditionally save training data, or load trained model and use it to predict traffic light state.
Note that currently the simulator is a few frames behind the state that it
provides so that on traffic light state transitions we still need to manually
correct the labels.csv file.

Add train.py to train a keras model and save it so that it can be used
for traffic light prediction.

Add test_model.py so we can run a quick sanity check on the model.